[MIDAS] fix type errors

+++ src/main/cc/mmio_catapult.cc
@@ -145,7 +145,7 @@ void tick() {
   top->clock = 1;
   top->eval();
 #if VM_TRACE
-  if (tfp) tfp->dump(main_time);
+  if (tfp) tfp->dump((double) main_time);
 #endif // VM_TRACE
   main_time++;
 
@@ -160,7 +160,7 @@ void tick() {
   top->clock = 0;
   top->eval();
 #if VM_TRACE
-  if (tfp) tfp->dump(main_time);
+  if (tfp) tfp->dump((double) main_time);
 #endif // VM_TRACE
   main_time++;
